fixing automod

+++ events/guild/messageCreate.js
@@ -39,6 +39,8 @@ module.exports = async (Discord, client, message) => {
         if (!message.content && message.attachments.size <= 0 && message.stickers.size <= 0) {
             if (!(message.member.roles.cache.has('614196214078111745')) && message.channel.id !== data.modChat) {
                 if (data.deletepoll === true) {
+                    if (message.channel.id === '1110289979663470714') return;
+                    
                     await message.reply({ content: 'Polls have been disallowed for posting!' }).then((m) => setTimeout(() => m.delete(), 4000));
                     await message.delete();
                 }
